Refactor: Point 패키지 네이밍 롤백
