+++ src/main/java/supernova/whokie/point_record/PointRecordOption.java
@@ -1,4 +1,4 @@
-package supernova.whokie.point;
+package supernova.whokie.point_record;
 
 public enum PointRecordOption {
     ALL, USED, CHARGED
